create actions.auth
Create actions.auth and reorganize login and updatePassword from actions.customers
